requirements.txt
accelerate==1.3.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-cors==0.7.0
aiosignal==1.3.2
airportsdata==20241001
annotated-types==0.7.0
antlr4-python3-runtime==4.7.2
anyio==4.8.0
appdirs==1.4.4
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
av==14.0.1
bitsandbytes==0.42.0
black==24.1.0
blake3==1.0.4
cachetools==5.5.1
certifi==2024.12.14
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorful==0.5.6
comm==0.2.2
compressed-tensors==0.9.1
conda-pack==0.8.1
contourpy==1.3.1
cupy-cuda12x==13.4.0
cycler==0.12.1
DataProperty==1.1.0
datasets==2.16.1
debugpy==1.8.12
decorator==5.1.1
decord==0.6.0
deepspeed==0.15.4
depyf==0.18.0
dill==0.3.7
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
email_validator==2.2.0
et_xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastrlock==0.8.3
ffmpy==0.5.0
filelock==3.17.0
fire==0.7.0
flash_attn==2.7.3
fonttools==4.55.6
frozenlist==1.5.0
fsspec==2023.10.0
ftfy==6.3.1
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.66.0
gradio==5.12.0
gradio_client==1.5.4
grpcio==1.70.0
h11==0.14.0
hf_transfer==0.1.9
hjson==3.1.0
httpcore==0.16.3
httptools==0.6.4
httpx==0.23.3
huggingface-hub==0.27.1
identify==2.6.6
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
inquirerpy==0.3.4
interegular==0.3.3
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isort==5.13.2
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.3
jiter==0.8.2
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lark==1.2.2
latex2sympy2==1.9.1
latex2sympy2_extended==1.0.6
liger_kernel==0.5.2
linkify-it-py==2.0.3
-e git+https://github.com/hiyouga/LLaMA-Factory.git@0f45982bac6b65533a94054ea5f792cb0f9e5a1f#egg=llamafactory
llvmlite==0.43.0
lm-format-enforcer==0.10.9
-e git+https://github.com/EvolvingLMMs-Lab/lmms-eval.git@e41110416177d60da3165ea73124a2cd5e48c53a#egg=lmms_eval
loguru==0.7.3
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
math-verify==0.5.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.15.0
mistral_common==1.5.1
mpmath==1.3.0
-e git+ssh://git@github.com/ssmisya/PRMBench.git@f0d47f51feb82496c21f059dc5e8d3789c09a00f#egg=mr_toolkit
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.3
nltk==3.9.1
nodeenv==1.9.1
numba==0.60.0
numexpr==2.10.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.59.9
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post5
pathspec==0.12.1
pathvalidate==3.2.3
peft==0.12.0
pexpect==4.9.0
pfzy==0.3.4
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.1.1
pre_commit==4.1.0
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==3.20.0
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyairports==2.1.1
pyarrow==19.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycocoevalcap==1.2
pycocotools==2.0.8
pycountry==24.6.1
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
pytablewriter==1.2.1
pytest==8.3.4
python-box==7.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qwen-vl-utils==0.0.9
r1-v @ file:///mnt/petrelfs/songmingyang/code/reasoning/R1-V/src/r1-v
ray==2.40.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rich-toolkit==0.13.2
rouge-chinese==1.0.3
rpds-py==0.22.3
rsa==4.9
ruff==0.9.3
sacrebleu==2.5.1
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
semantic-version==2.10.0
sentence-transformers==3.4.0
sentencepiece==0.2.0
sentry-sdk==2.20.0
setproctitle==1.3.4
shellingham==1.5.4
shtab==1.7.1
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sqlitedict==2.1.0
sse-starlette==2.2.1
stack-data==0.6.3
-e git+ssh://git@github.com/Kuvvius/stare_open.git@8710dca0304878934e5c757c0ef29b649754f2f1#egg=stare_open
starlette==0.41.3
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tenacity==8.3.0
termcolor==2.5.0
textual==1.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.14
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.13.2
-e git+ssh://git@github.com/zhaochen0110/tool-agent.git@5b5fa8c6ab633b375ab29e146a7aec0dcb401b4e#egg=tool_server
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers==4.49.0
transformers-stream-generator==0.0.5
triton==3.2.0
trl==0.14.0
typepy==1.3.4
typer==0.15.1
typing_extensions==4.12.2
tyro==0.8.14
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.1
vllm==0.7.3
wandb==0.16.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
widgetsnbextension==4.0.13
wrapt==1.17.2
-e git+https://github.com/facebookresearch/xformers.git@c5841688d00358b1ddd583f217231aeced66e01c#egg=xformers
xgrammar @ file:///mnt/petrelfs/songmingyang/softwares/package/xgrammar/python
xxhash==3.5.0
yarl==1.18.3
yt-dlp==2025.1.26
zipp==3.21.0
zss==1.2.0
zstandard==0.23.0
